ConsoleLogger for REPL mode logging
Start of a ConsoleLogger optimized for viewing log records in the REPL.
This is the SimpleLogger code copied and made aware of display size so
that some information can be right justified.

Also make this the default logger type so that users can get interface
improvements by upgrading the stdlib Logging library rather than being
bound to Base.